+++ index-test.php
@@ -10,7 +10,7 @@ if (!in_array(@$_SERVER['REMOTE_ADDR'], ['127.0.0.1', '::1'])) {
     die('You are not allowed to access this file.');
 }
 
-defined('YII_DEBUG') or define('YII_DEBUG', true);
+/*defined('YII_DEBUG') or define('YII_DEBUG', true);*/
 defined('YII_ENV') or define('YII_ENV', 'test');
 defined('YII_ENV_TEST') or define('YII_ENV_TEST', true);