+++ protected/humhub/modules/admin/messages/de/notifications_AdminNotificationCategory.php
@@ -1,5 +1,5 @@
 <?php
 return array (
-  'Administrative' => 'Administrative',
-  'Receive Notifications for administrative events like available updates.' => 'Erhalte Benachrichtungen für Ereignisse aus dem Bereich der Administragion (z.B. verfügbare Updates)',
+  'Administrative' => 'Administrativ',
+  'Receive Notifications for administrative events like available updates.' => 'Erhalte Benachrichtungen für Ereignisse aus dem Bereich der Administration (z.B. verfügbare Updates)',
 );