+++ protected/humhub/modules/admin/messages/nl/forms_LogsSettingsForm.php
@@ -1,10 +1,10 @@
 <?php
 return array (
-  '1 month' => '',
-  '1 week' => '',
-  '1 year' => '',
-  '2 weeks' => '',
-  '3 months' => '',
-  '6 months' => '',
+  '1 month' => '1 maand',
+  '1 week' => '1 week',
+  '1 year' => '1 jaar',
+  '2 weeks' => 'twee weken',
+  '3 months' => 'drie maanden',
+  '6 months' => 'zes maanden',
   'never' => 'nooit',
 );