+++ protected/humhub/modules/content/messages/nl/widgets_views_deleteLink.php
@@ -3,5 +3,5 @@ return array (
   '<strong>Confirm</strong> post deletion' => '<strong>Bevestig</strong> bericht verwijderen',
   'Cancel' => 'Annuleren',
   'Delete' => 'Verwijder',
-  'Do you really want to delete this post? All likes and comments will be lost!' => 'Ben je zeker dat je dit bericht wilt verwijderen? Alle likes en reacties zullen verloren gaan!',
+  'Do you really want to delete this post? All likes and comments will be lost!' => 'Weet u zeker dat u dit bericht wilt verwijderen? Alle vind-ik-leuks en reacties zullen verloren gaan!',
 );