+++ protected/humhub/modules/file/messages/fr/widgets_views_fileUploadButton.php
@@ -1,4 +1,4 @@
 <?php
 return array (
-  'Upload files' => 'Envoyer des fichiers',
+  'Upload files' => 'Transférer les fichiers',
 );