+++ protected/humhub/modules/installer/messages/nl/controllers_ConfigController.php
@@ -1,10 +1,10 @@
 <?php
 return array (
-  'Calvin Klein – Between love and madness lies obsession.' => '',
-  'Create Admin Account' => 'Maak Admin Account',
-  'Nike – Just buy it. ;Wink;' => '',
-  'We\'re looking for great slogans of famous brands. Maybe you can come up with some samples?' => '',
-  'Welcome Space' => 'Welkom Space',
-  'Yay! I\'ve just installed HumHub ;Cool;' => 'JEEJ! Ik heb net HumHub geïnstalleerd ;Cool;',
-  'Your first sample space to discover the platform.' => 'Je eerste voorbeeld space om het platform te ontdekken.',
+  'Calvin Klein – Between love and madness lies obsession.' => 'Calvin Klein – Between love and madness lies obsession.',
+  'Create Admin Account' => 'Maak beheerdersaccount',
+  'Nike – Just buy it. ;Wink;' => 'Nike – Just buy it. ;Wink;',
+  'We\'re looking for great slogans of famous brands. Maybe you can come up with some samples?' => 'We zoeken briljante leuzen van bekende merken.  Kunt u enkele voorbeelden bedenken?',
+  'Welcome Space' => 'Welkom Ruimte',
+  'Yay! I\'ve just installed HumHub ;Cool;' => 'JEEJ! Ik heb net HumHub geïnstalleerd ;Gaaf;',
+  'Your first sample space to discover the platform.' => 'Uw eerste voorbeeld ruimte om het platform te ontdekken.',
 );