+++ protected/humhub/modules/notification/messages/nl/models_forms_NotificationSettings.php
@@ -1,6 +1,6 @@
 <?php
 return array (
-  'Allow desktop notifications by default.' => '',
-  'Receive \'New Content\' Notifications for the following spaces' => '',
-  'Receive desktop notifications when you are online.' => 'Ontvang desktop notificaties wanner je online bent',
+  'Allow desktop notifications by default.' => 'Sta bureaublad meldingen standaard toe.',
+  'Receive \'New Content\' Notifications for the following spaces' => 'Ontvang \'Nieuwe inhoud\' meldingen voor de volgende ruimten',
+  'Receive desktop notifications when you are online.' => 'Ontvang bureaublad meldingen wanneer u online bent',
 );