+++ protected/humhub/modules/notification/messages/nl/widgets_views_list.php
@@ -1,7 +1,7 @@
 <?php
 return array (
   'Mark all as seen' => 'Markeer alles als gezien',
-  'Notifications' => 'Notificaties',
-  'Show all notifications' => 'Alle notificaties weergeven',
-  'There are no notifications yet.' => 'Er zijn nog geen notificaties.',
+  'Notifications' => 'Meldingen',
+  'Show all notifications' => 'Alle meldingen weergeven',
+  'There are no notifications yet.' => 'Er zijn nog geen meldingen.',
 );