+++ protected/humhub/modules/space/messages/fr/notifications_FriendshipNotificationCategory.php
@@ -1,5 +1,5 @@
 <?php
 return array (
   'Friendship' => 'Amitié',
-  'Receive Notifications for Friendship Request and Approval events.' => '',
+  'Receive Notifications for Friendship Request and Approval events.' => 'Recevoir des alertes pour les demandes d\'amitié et les acceptations.',
 );