+++ protected/humhub/modules/space/messages/fr/widgets_spaceChooserItem.php
@@ -1,6 +1,6 @@
 <?php
 return array (
-  'This space is archived' => '',
+  'This space is archived' => 'Cet espace est archivé',
   'You are a member of this space' => 'Vous êtes membre de cet espace',
-  'You are following this space' => '',
+  'You are following this space' => 'Vous suivez cet espace',
 );