+++ protected/humhub/modules/space/messages/nl/widgets_spaceChooserItem.php
@@ -1,6 +1,6 @@
 <?php
 return array (
-  'This space is archived' => '',
-  'You are a member of this space' => 'Je bent lid van deze ruimte',
-  'You are following this space' => '',
+  'This space is archived' => 'Deze ruimte is gearchiveerd.',
+  'You are a member of this space' => 'U bent lid van deze ruimte',
+  'You are following this space' => 'U volgt deze ruimte.',
 );